requirements.txt
clip
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-cors==0.8.1
aiosignal==1.3.2
airportsdata==20250622
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
anykeystore==0.2
astor==0.8.1
attrs==25.3.0
blake3==1.0.5
braceexpand==0.1.7
# brotlicffi @ file:///croot/brotlicffi_1736182461069/work
cachetools==5.5.2
# certifi @ file:///croot/certifi_1750863846190/work/certifi
# cffi @ file:///croot/cffi_1736182485317/work
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
coloredlogs==15.0.1
colorful==0.5.7
compressed-tensors==0.9.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.2
datasets==3.6.0
defusedxml==0.7.1
depyf==0.18.0
diffusers==0.34.0
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
docker-pycreds==0.4.0
easydict==1.13
einops==0.8.1
fastapi==0.115.14
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.3.0
gguf==0.10.0
gitdb==4.0.12
GitPython==3.1.44
# gmpy2 @ file:///croot/gmpy2_1738085463648/work
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.2
humanfriendly==10.0
hupper==1.12.1
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
insightface==0.7.3
interegular==0.3.3
jaxtyping==0.3.2
# Jinja2 @ file:///croot/jinja2_1741710844255/work
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lark==1.2.2
lazy_loader==0.4
lightning==2.5.2
lightning-utilities==0.14.3
lm-format-enforcer==0.10.11
# MarkupSafe @ file:///croot/markupsafe_1738584038848/work
matplotlib==3.10.3
mistral_common==1.6.3
mkl-service==2.4.0
# mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
# mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
# mpmath @ file:///croot/mpmath_1690848262763/work
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.3
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.5
ninja==1.11.1.4
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.3.1
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.22.0
openai==1.93.0
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.11.0.86
opentelemetry-api==1.34.1
opentelemetry-exporter-prometheus==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
outlines==0.1.11
outlines_core==0.1.26
packaging==25.0
pandas==2.3.0
partial-json-parser==0.2.1.1.post6
PasteDeploy==3.1.0
peft==0.16.0
pillow==11.3.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs==4.3.8
pluggy==1.6.0
prettytable==3.16.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
propcache==0.3.2
proto-plus==1.26.1
# protobuf==4.25.8
psutil==7.0.0
py==1.11.0
py-cpuinfo==9.0.0
py-spy==0.4.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
# pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
pyramid==2.0.2
# PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python3-openid==3.2.0
pytorch-lightning==2.5.2
pytz==2025.2
# PyYAML @ file:///croot/pyyaml_1728657952215/work
pyzmq==27.0.0
ray==2.47.1
referencing==0.36.2
regex==2024.11.6
# requests @ file:///croot/requests_1750426329888/work
requests-oauthlib==2.0.0
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.16.0
sentencepiece==0.2.0
sentry-sdk==2.32.0
setproctitle==1.3.6
simsimd==6.4.9
six==1.17.0
smart_open==7.3.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
stringzilla==3.12.5
# sympy @ file:///croot/sympy_1738108488918/work
threadpoolctl==3.6.0
tifffile==2025.6.11
tiktoken==0.9.0
tokenizers==0.21.2
# torch==2.7.1
torch-fidelity==0.3.0
# torchaudio==2.7.1
torchmetrics==1.4.0.post0
# torchvision==0.22.1
tox==3.28.0
tqdm==4.66.4
traitlets==5.14.3
transaction==4.0
transformers==4.53.1
translationstring==1.4
triton==3.3.1
trl==0.19.0
typeguard==2.13.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2024.1
urllib3==1.26.18
uvicorn==0.35.0
uvloop==0.21.0
velruse==1.1.1
venusian==3.1.0
virtualenv==20.26.2
wadler_lindig==0.1.7
wandb==0.21.0
watchdog==4.0.1
watchfiles==1.1.0
wcwidth==0.2.13
webdataset==0.2.86
WebOb==1.8.7
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
wtforms-recaptcha==0.3.2
xformers==0.0.31
xgrammar==0.1.20
xxhash==3.4.1
yarl==1.20.1
zipp==3.23.0
zope.deprecation==5.0
zope.interface==6.4.post2
zope.sqlalchemy==3.1

